Migration downloads individual migration binaries
When performing a migration after installing a new ipfs version, ipfs-update finds the necessary mirgations, downloads the latest version of them from the distribution site, unpacks the executables, and runs the migrations in order.  It uses the code located in go-ipfs/repo/fsrepo/migrations to do this.

Also used go-ipfs/repo/fsrepo/migrations
- fetching binaries for things other than migrations
- fetching version information about binaries other than migrations
- looking at current repo version
- checking for local ipfs api availability